scripts/requirements.txt
pandas
tabula-py
pdfplumber  # for text-based PDF extraction
PyMuPDF      # alternative PDF parser
requests